Move colorname stuff to own file, add fav combo tooltips
