+++ src/base/zdefs.cpp
@@ -1781,3 +1781,21 @@ size_and_pos::size_and_pos(int32_t nx, int32_t ny, int32_t nw, int32_t nh, int32
 	: x(nx), y(ny), w(nw), h(nh), xscale(xsc), yscale(ysc)
 {}
 
+int newcombo::each_tile(std::function<bool(int32_t)> proc) const
+{
+	int tile = o_tile;
+	int frame = 0;
+	do
+	{
+		if(proc(tile))
+			return frame;
+		if(++frame >= frames)
+			break;
+		tile += ((1+skipanim)*frame);
+		if(int rowoffset = TILEROW(tile)-TILEROW(o_tile))
+			tile += skipanimy * rowoffset * TILES_PER_ROW;
+	}
+	while(true);
+	return -1;
+}
+
